# untier-server/src/main/resources/openapiv3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Untier API",
    "version": "0.1.0"
  },
  "tags": [
    {
      "name": "AuthoringService"
    },
    {
      "name": "QueryService"
    }
  ],
  "paths": {
    "/postings": {
      "get": {
        "tags": [
          "QueryService"
        ],
        "summary": "getPostings",
        "operationId": "getPostings",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Posting"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AuthoringService"
        ],
        "summary": "publish",
        "operationId": "publish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPosting"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "maxLength": 400
                  },
                  "attachment": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["text", "attachment"]
              },
              "encoding": {
                "attachment": {
                  "contentType": "image/gif, image/jpeg, image/png, image/webp"
                }
              }
            },
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "400": {
            "description": "failed operation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "failed authentication"
          }
        },
        "security": [
          {
            "jwt": [
              "profile",
              "email",
              "offline_access"
            ]
          }
        ]
      }
    },
    "/postings/{id}": {
      "get": {
        "tags": [
          "QueryService"
        ],
        "summary": "getPosting",
        "operationId": "getPosting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Posting"
                }
              }
            }
          },
          "404": {
            "description": "not found"
          }
        }
      },
      "delete": {
        "tags": [
          "AuthoringService"
        ],
        "summary": "retract",
        "operationId": "retract",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation"
          },
          "403": {
            "description": "failed authentication"
          }
        },
        "security": [
          {
            "jwt": [
              "profile",
              "email",
              "offline_access"
            ]
          }
        ]
      }
    },
    "/attachments/{id}": {
      "get": {
        "tags": [
          "QueryService"
        ],
        "summary": "getAttachment",
        "operationId": "getAttachment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Attachment": {
        "type": "object",
        "properties": {
          "bytes": {
            "type": "string",
            "format": "byte"
          },
          "mediaType": {
            "type": "string",
            "maxLength": 255
          }
        },
        "required": [
          "bytes",
          "mediaType"
        ]
      },
      "DraftPosting": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 400
          },
          "attachment": {
            "$ref": "#/components/schemas/Attachment"
          }
        },
        "required": [
          "text"
        ]
      },
      "Posting": {
        "type": "object",
        "properties": {
          "author": {
            "type": "string",
            "maxLength": 255
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "text": {
            "type": "string",
            "maxLength": 400
          },
          "published": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "author",
          "id",
          "text",
          "published"
        ]
      },
      "Problem": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "title"
        ]
      }
    },
    "securitySchemes": {
      "jwt": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://localhost:1337/auth",
            "scopes": {
              "profile": "",
              "email": "",
              "offline_access": ""
            }
          },
          "authorizationCode": {
            "authorizationUrl": "http://localhost:1337/auth",
            "tokenUrl": "http://localhost:1337/token",
            "refreshUrl": "http://localhost:1337/token",
            "scopes": {
              "profile": "",
              "email": "",
              "offline_access": ""
            }
          }
        }
      }
    }
  }
}